% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectionRecoverBlocking.R
\name{DetectionRecoverBlocking}
\alias{DetectionRecoverBlocking}
\title{Title Function used to detect edge and denoise image with blocking capability}
\usage{
DetectionRecoverBlocking(z, h, k, k1, block_size = 64)
}
\arguments{
\item{z}{Input image, single channel only.}

\item{h}{Parameter to adjust size of neighborhood during least square calculation.}

\item{k}{Tuning parameter, size of neighborhood for edge calculation.}

\item{k1}{Tuning parameter, size of neighborhood for denoising.}

\item{block_size}{Size of blocks you wish to use, default size 64x64.}
}
\value{
A list of detected edges (an 0, 1 matrix of edge locations with 1 being edge location) and denoised image.
}
\description{
Detect edges and recover images in one step, efficiently handling non-square images by dividing them into smaller blocks to speed up calculations.
}
