% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_detect.R
\name{edge_detect}
\alias{edge_detect}
\title{Function to detect edges/jumps on SQUARE image, i.e. equal width and height}
\usage{
edge_detect(z, k, h, alpha_n = 0.05)
}
\arguments{
\item{z}{Input image, single channel only.}

\item{k}{Tuning parameter, size of neighborhood for edge calculation, should be an odd integer.}

\item{h}{Parameter to adjust size of neighborhood during least square calculation, should be an integer.}

\item{alpha_n}{Parameter to adjust detection sensitivity, default 0.05.}
}
\value{
An 0, 1 matrix of edge locations with 1 being edge location
}
\description{
Function to detect edges/jumps on SQUARE image, i.e. equal width and height
}
\references{
Qiu, Peihua. “Discontinuous Regression Surfaces Fitting.” The Annals of Statistics 26, no. 6 (1998): 2218–45.
}
