% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump_preserving_estimator.R
\name{jump_preserving_estimator}
\alias{jump_preserving_estimator}
\title{Generate a jump-preserving surface estimator from noisy data.}
\usage{
jump_preserving_estimator(z, h, sigma_hat, alpha = 0.05)
}
\arguments{
\item{z}{Input surface, should be single channel image}

\item{h}{Parameter to adjust size of neighborhood for local surface estimation, should be an integer.}

\item{sigma_hat}{Estimated noise level (noise variance) of input surface}

\item{alpha}{Significance level to control the reconstruction performance of jump, default 0.05}
}
\value{
Estimated surface.
}
\description{
Generate a jump-preserving surface estimator from noisy data.
}
\references{
Qiu, Peihua. “The Local Piecewisely Linear Kernel Smoothing Procedure for Fitting Jump Regression Surfaces.” Technometrics 46, no. 1 (February 1, 2004): 87–98. https://doi.org/10.1198/004017004000000149.

Qiu, Peihua. “Jump-Preserving Surface Reconstruction from Noisy Data.” Annals of the Institute of Statistical Mathematics 61, no. 3 (September 1, 2009): 715–51. https://doi.org/10.1007/s10463-007-0166-9.
}
