% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_surface.R
\name{recover_surface}
\alias{recover_surface}
\title{Function to recover noisy surface with jump detail preserved.}
\usage{
recover_surface(z, kk, edge)
}
\arguments{
\item{edge}{}
}
\description{
Function to recover noisy surface with jump detail preserved.
}
